requirements.txt
matplotlib==3.7.2
scikit-spatial==6.0.0
scipy==1.4.1
tqdm==4.66.1
tensorflow==2.2.0
tensorflow-gpu==2.2.0
waymo-open-dataset-tf-2-2-0==1.3.1